Human moves after Idle to Walk animation is over + Fixed walking animations
added the "HumanBeginWalkIsPlaying" bool parameter for the Human animator. A script on HumanBeginWalk turns it to true when the animation starts and flase when it ends.
The Human Follow Script will perform transform manipulation only when HumanBeginWalkIsPlaying is false.

Fixed the walk cycle animation and walk begin animation to have the same ground line. A better transition between them would be nice to have